Add files via upload
The project seeks to go further and use clustering, hypothesis testing as well as attribute independence analysis to delve deeper into the world development indicators such as birth rate, infant mortality, business taxes rates etc., in a bid to pinpoint key relationships in the underlying socio-economic environment of the world today. Regional classifications, based on socio-economic development indicators, show that there are groupings of countries with similar development characteristics and this makes it easier to review, and analyze regional disparities. Hypothesis testing allows us to determine if there is indeed a correlation, such as between the business tax rates and GDP growth or spending on health care and life expectancy. Also, the interdependence and the extent of the impact of these factors may be determined by identifying how precisely each component correlates with the others. And comparative analysis of these indicators were performed using the clustering techniques such as K means, hierarchical clustering, DBSCAN and Spectral. The silhouette score for the clusters using K means and Hierarchical clustering shows up better analysis of the development status of the countries based on the world development features.